Make windows-specific projects restore for windows
Many projects were using windows-specific packages.  These projects
need to restore for windows specifically to avoid compat errors during
restore.
