+++ pkg/redist/System.Runtime.InteropServices/lib/project.json
@@ -5,5 +5,9 @@
   },
   "frameworks": {
     "dnxcore50": {}
+  },
+  "runtimes": {
+    "win7-x86": {},
+    "win7-x64": {}
   }
 }
