+++ src/Microsoft.Win32.Registry.AccessControl/tests/project.json
@@ -16,5 +16,9 @@
     "dnxcore50": {
       "imports": "portable-net45+win8"
     }
+  },
+  "runtimes": {
+    "win7-x86": {},
+    "win7-x64": {}
   }
 }
