+++ src/Microsoft.Win32.Registry/tests/project.json
@@ -14,5 +14,9 @@
     "dnxcore50": {
       "imports": "portable-net45+win8"
     }
+  },
+  "runtimes": {
+    "win7-x86": {},
+    "win7-x64": {}
   }
 }
