+++ src/System.Diagnostics.Process/tests/ProcessStartInfoTests.cs
@@ -15,7 +15,7 @@ using System.Text;
 
 namespace System.Diagnostics.Tests
 {
-    public class ProcessStartInfoTests : ProcessTestBase
+    public partial class ProcessStartInfoTests : ProcessTestBase
     {
         [Fact]
         public void TestEnvironmentProperty()
@@ -313,59 +313,6 @@ namespace System.Diagnostics.Tests
             }
         }
 
-        [Fact, PlatformSpecific(PlatformID.Windows), OuterLoop] // Requires admin privileges
-        public void TestUserCredentialsPropertiesOnWindows()
-        {
-            string username = "test", password = "PassWord123!!";
-            try
-            {
-                Interop.NetUserAdd(username, password);
-            }
-            catch (Exception exc)
-            {
-                Console.Error.WriteLine("TestUserCredentialsPropertiesOnWindows: NetUserAdd failed: {0}", exc.Message);
-                return; // test is irrelevant if we can't add a user
-            }
-
-            Process p = CreateProcessLong();
-
-            p.StartInfo.LoadUserProfile = true;
-            p.StartInfo.UserName = username;
-            p.StartInfo.PasswordInClearText = password;
-
-            SafeProcessHandle handle = null;
-            try
-            {
-                p.Start();
-                if (Interop.OpenProcessToken(p.SafeHandle, 0x8u, out handle))
-                {
-                    SecurityIdentifier sid;
-                    if (Interop.ProcessTokenToSid(handle, out sid))
-                    {
-                        string actualUserName = sid.Translate(typeof(NTAccount)).ToString();
-                        int indexOfDomain = actualUserName.IndexOf('\\');
-                        if (indexOfDomain != -1)
-                            actualUserName = actualUserName.Substring(indexOfDomain + 1);
-
-                        bool isProfileLoaded = GetNamesOfUserProfiles().Any(profile => profile.Equals(username));
-
-                        Assert.Equal(username, actualUserName);
-                        Assert.True(isProfileLoaded);
-                    }
-                }
-            }
-            finally
-            {
-                if (handle != null)
-                    handle.Dispose();
-
-                if (!p.HasExited)
-                    p.Kill();
-
-                Interop.NetUserDel(null, username);
-                Assert.True(p.WaitForExit(WaitInMS));
-            }
-        }
 
         [Fact, PlatformSpecific(PlatformID.AnyUnix)]
         public void TestUserCredentialsPropertiesOnUnix()
@@ -397,30 +344,6 @@ namespace System.Diagnostics.Tests
 
                 Assert.True(p.WaitForExit(WaitInMS));
             }
-        }
-
-        private static List<string> GetNamesOfUserProfiles()
-        {
-            List<string> userNames = new List<string>();
-
-            string[] names = Registry.Users.GetSubKeyNames();
-            for (int i = 1; i < names.Length; i++)
-            {
-                try
-                {
-                    SecurityIdentifier sid = new SecurityIdentifier(names[i]);
-                    string userName = sid.Translate(typeof(NTAccount)).ToString();
-                    int indexofDomain = userName.IndexOf('\\');
-                    if (indexofDomain != -1)
-                    {
-                        userName = userName.Substring(indexofDomain + 1);
-                        userNames.Add(userName);
-                    }
-                }
-                catch (Exception) { }
-            }
-
-            return userNames;
         }
     }
 }
