+++ src/System.Diagnostics.Process/tests/ProcessStartInfoTests.windows.cs
@@ -1,0 +1,98 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using Microsoft.Win32;
+using Microsoft.Win32.SafeHandles;
+using System.Collections;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Runtime.InteropServices;
+using System.Security.Principal;
+using Xunit;
+using System.Text;
+
+namespace System.Diagnostics.Tests
+{
+    public partial class ProcessStartInfoTests : ProcessTestBase
+    {
+        [Fact, PlatformSpecific(PlatformID.Windows), OuterLoop] // Requires admin privileges
+        public void TestUserCredentialsPropertiesOnWindows()
+        {
+            string username = "test", password = "PassWord123!!";
+            try
+            {
+                Interop.NetUserAdd(username, password);
+            }
+            catch (Exception exc)
+            {
+                Console.Error.WriteLine("TestUserCredentialsPropertiesOnWindows: NetUserAdd failed: {0}", exc.Message);
+                return; // test is irrelevant if we can't add a user
+            }
+
+            Process p = CreateProcessLong();
+
+            p.StartInfo.LoadUserProfile = true;
+            p.StartInfo.UserName = username;
+            p.StartInfo.PasswordInClearText = password;
+
+            SafeProcessHandle handle = null;
+            try
+            {
+                p.Start();
+                if (Interop.OpenProcessToken(p.SafeHandle, 0x8u, out handle))
+                {
+                    SecurityIdentifier sid;
+                    if (Interop.ProcessTokenToSid(handle, out sid))
+                    {
+                        string actualUserName = sid.Translate(typeof(NTAccount)).ToString();
+                        int indexOfDomain = actualUserName.IndexOf('\\');
+                        if (indexOfDomain != -1)
+                            actualUserName = actualUserName.Substring(indexOfDomain + 1);
+
+                        bool isProfileLoaded = GetNamesOfUserProfiles().Any(profile => profile.Equals(username));
+
+                        Assert.Equal(username, actualUserName);
+                        Assert.True(isProfileLoaded);
+                    }
+                }
+            }
+            finally
+            {
+                if (handle != null)
+                    handle.Dispose();
+
+                if (!p.HasExited)
+                    p.Kill();
+
+                Interop.NetUserDel(null, username);
+                Assert.True(p.WaitForExit(WaitInMS));
+            }
+        }
+
+        private static List<string> GetNamesOfUserProfiles()
+        {
+            List<string> userNames = new List<string>();
+
+            string[] names = Registry.Users.GetSubKeyNames();
+            for (int i = 1; i < names.Length; i++)
+            {
+                try
+                {
+                    SecurityIdentifier sid = new SecurityIdentifier(names[i]);
+                    string userName = sid.Translate(typeof(NTAccount)).ToString();
+                    int indexofDomain = userName.IndexOf('\\');
+                    if (indexofDomain != -1)
+                    {
+                        userName = userName.Substring(indexofDomain + 1);
+                        userNames.Add(userName);
+                    }
+                }
+                catch (Exception) { }
+            }
+
+            return userNames;
+        }
+    }
+}
