+++ src/System.Diagnostics.Process/tests/win/project.json
@@ -3,6 +3,7 @@
     "Microsoft.DotNet.xunit.performance": "1.0.0-alpha-build0029",
     "Microsoft.NETCore.Platforms": "1.0.1-rc3-23908",
     "Microsoft.Win32.Primitives": "4.0.0",
+    "Microsoft.Win32.Registry": "4.0.0-rc3-23908",
     "System.Collections": "4.0.10",
     "System.Console": "4.0.0-rc3-23908",
     "System.IO": "4.0.10",
@@ -27,5 +28,9 @@
     "dnxcore50": {
       "imports": "portable-net45+win8"
     }
+  },
+  "runtimes": {
+    "win7-x86": {},
+    "win7-x64": {}
   }
 }
