+++ src/System.IO.FileSystem.AccessControl/tests/project.json
@@ -22,5 +22,9 @@
     "dnxcore50": {
       "imports": "portable-net45+win8"
     }
+  },
+  "runtimes": {
+    "win7-x86": {},
+    "win7-x64": {}
   }
 }
