+++ src/System.IO.Pipes/src/project.json
@@ -15,7 +15,6 @@
         "System.Runtime.InteropServices": "4.0.20",
         "System.Security.Principal": "4.0.0",
         "System.Threading": "4.0.10",
-        "System.Threading.Overlapped": "4.0.0",
         "System.Threading.Tasks": "4.0.10"
       }
     },
