+++ src/System.IO.Pipes/src/win/project.json
@@ -3,7 +3,6 @@
     "dnxcore50": {
       "dependencies": {
         "Microsoft.NETCore.Platforms": "1.0.1-rc3-23908",
-        "System.Collections": "4.0.10",
         "System.Diagnostics.Contracts": "4.0.0",
         "System.Diagnostics.Debug": "4.0.10",
         "System.Diagnostics.Tools": "4.0.0",
@@ -14,16 +13,15 @@
         "System.Runtime.Extensions": "4.0.10",
         "System.Runtime.Handles": "4.0.0",
         "System.Runtime.InteropServices": "4.0.20",
-        "System.Text.Encoding": "4.0.10",
-        "System.Text.Encoding.Extensions": "4.0.10",
+        "System.Security.Principal": "4.0.0",
         "System.Threading": "4.0.10",
+        "System.Threading.Overlapped": "4.0.0",
         "System.Threading.Tasks": "4.0.10"
       }
-    },
-    "net46": {
-      "dependencies": {
-        "Microsoft.TargetingPack.NETFramework.v4.6": "1.0.1"
-      }
     }
+  },
+  "runtimes": {
+    "win7-x86": {},
+    "win7-x64": {}
   }
 }
