+++ src/System.IO.Pipes/tests/project.json
@@ -11,7 +11,6 @@
     "System.Runtime.InteropServices": "4.0.20",
     "System.Runtime.InteropServices.RuntimeInformation": "4.0.0-rc3-23908",
     "System.Security.Principal": "4.0.0",
-    "System.Threading.Overlapped": "4.0.0",
     "System.Threading.Tasks": "4.0.10",
     "xunit": "2.1.0",
     "xunit.netcore.extensions": "1.0.0-prerelease-00187"
