+++ src/System.Threading.AccessControl/tests/project.json
@@ -17,5 +17,9 @@
     "dnxcore50": {
       "imports": "portable-net45+win8"
     }
+  },
+  "runtimes": {
+    "win7-x86": {},
+    "win7-x64": {}
   }
 }
