+++ src/System.Threading.Overlapped/tests/project.json
@@ -9,5 +9,9 @@
     "dnxcore50": {
       "imports": "portable-net45+win8"
     }
+  },
+  "runtimes": {
+    "win7-x86": {},
+    "win7-x64": {}
   }
 }
